+++ src/main/java/com/refinedmods/refinedstorage/RS.java
@@ -48,7 +48,6 @@ public final class RS {
 
         RSBlocks.register();
         RSItems.register();
-        RSLootFunctions.register();
 
         FMLJavaModLoadingContext.get().getModEventBus().addListener(CommonSetup::onCommonSetup);
         FMLJavaModLoadingContext.get().getModEventBus().addGenericListener(BlockEntityType.class, CommonSetup::onRegisterBlockEntities);